requirements.txt
# Minimal requirements for Snowflake-native validation framework
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
snowflake-connector-python>=3.0.0
plotly>=5.18.0
requests>=2.28.0
pyyaml>=6.0.0
